+++ src/sweeper.h
@@ -19,17 +19,17 @@ limitations under the License.
 
 #include <vector>
 
-#include "minimalloc.h"
 #include "absl/container/btree_set.h"
 #include "absl/container/flat_hash_set.h"
+#include "minimalloc.h"
 
 namespace minimalloc {
 
 // An index into a list of schedule "cross sections."
-using SectionIdx = int;
+using SectionIdx = int64_t;
 
 // A count of buffers crossing between adjacent sections.
-using CutCount = int;
+using CutCount = int64_t;
 
 // An interval that defines a subset of ranges (e.g., where a buffer is
 // continuously active).
